x86 JIT compiler!
Finally I've made a JIT compiler for SSVM :)
It's really primitive and compiles only a few opcodes (so not the full
set, yet), but works! ;)
As I said: for now, it only compiles a few instructions
(mov/add/sub/mul/div/if_*/jmp/tjmp/fjmp/call/icall/stop/nop), so it
won't properly compile most SScript programs, but nevertheless: it's
something! Also, the generated output x86 code is... em, ugly... ugly
and inefficient, but as this JIT compiler is developped, it'll slowly
become better and better. I think so.